Conflicts:
	chessaint/include/astar.h
	chessaint/include/graph.h
	chessaint/include/loopuci.h
	chessaint/include/uci.h
	chessaint/src/astar.c
	chessaint/src/graph.c
	chessaint/src/loopuci.c
	chessaint/src/main.c
	tests/unit/src/cunit_astar.c

(Conflicts are stupid problems comming from svn for most of
the files)

+++ chessaint/src/identifier.c
@@ -94,7 +94,7 @@ void stack_push(Stack *s, int item) {
   mpz_mul_si(*s, tmp, MAX_PAIRS);
 
   mpz_set(tmp, *s);
-  mpz_add_ui(*s, tmp, (unsigned long int) item);
+  mpz_add_ui(*s, tmp, item);
   mpz_clear(tmp);
 }
